+++ OAuth/Sources/PianoIDFormInfo.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+@objcMembers
+public class PianoIDFormInfo: NSObject {
+    
+    public let hasAllCustomFieldValuesFilled: Bool
+    
+    internal init(hasAllCustomFieldValuesFilled: Bool) {
+        self.hasAllCustomFieldValuesFilled = hasAllCustomFieldValuesFilled
+    }
+}